feature: introduzindo interface para os beans; validacao de possiveis status
