+++ src/main/java/com/github/viniciusmartins/qrcode/entity/enums/StatusEnum.java
@@ -1,5 +1,7 @@
 package com.github.viniciusmartins.qrcode.entity.enums;
 
 public enum StatusEnum {
-    OPEN
+    OPEN,
+    PAID,
+    CANCELED
 }
