+++ src/main/resources/messages.properties
@@ -1,9 +1,9 @@
 
 required.field=Field {0} is required
-invalid.field=Field {0} has a invalid format
 invalid.field.numeric.format=Numeric field {0} has a invalid format. Try value like 123.45
 invalid.field.numeric.range=Numeric field {0} should be greater then Zero
 invalid.field.date.format=Date field {0} has a invalid format. Try pattern yyyy-MM-dd
 invalid.field.date.range=Date field {0} should be a data from the future
+invalid.field.status=Status doesn't match the expected values
 entity.already.registered={0} already registered
 internal.server.error=Internal server error
